Finished Project (for now)
Refined and reconfigured calculation for QB and RB points, removing the base expected point value entirely, meaning all the positions project starting at 0 points. It's not flawless, obviously, but it's pretty accurate to differentiate between good/bad fantasy players. If I return to this, I'd add a way to handle rookies (although right now, they work, because it's been long enough since the start of the season). WRs tend to be overestimated, but that's about it. Super fun project to make, learning how to turn data into SQL databses and use them to make predictions about the future.